% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrApplyParamset.R
\name{getRemainingCombos}
\alias{getRemainingCombos}
\title{getRemainingParamsets generates the full set of paramset combos and
    removes the paramsets already processed, then returns a dataframe of
    remaining paramsets to be processed}
\usage{
getRemainingCombos(customParamsets = NULL, strategy, paramsetLabel,
  processedComboNums = NULL)
}
\arguments{
\item{strategy}{a name of a strategy or strategy object}
}
\description{
getRemainingParamsets generates the full set of paramset combos and
    removes the paramsets already processed, then returns a dataframe of
    remaining paramsets to be processed
}

