% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphParamsets.R
\name{tradeGraphs_asp}
\alias{tradeGraphsExt}
\alias{tradeGraphs_asp}
\title{graph applyParamset() output}
\usage{
tradeGraphs_asp(stats, free.params, params.filter = NULL, statistics,
  title = NULL, print_gr = FALSE, debug = FALSE, colsch = c("viridis",
  "heat", "rbow", "rbowStripesD1", "rbowStripesD2"), contour = FALSE,
  zcuts = 30, type = c("lp", "levelplot", "3d", "rgl", "hm", "heatmap"))
}
\arguments{
\item{colsch}{color scheme (defaults to 'viridis')}

\item{zcuts}{the number of distinct colors in the palette}
}
\description{
graph applyParamset() output
}

