% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeStatsRoll.R
\name{tradeStatsRoll}
\alias{tradeStatsRoll}
\title{Produces 'rolling' statistics based on the 'period' (unit size),
span (input data window size), and step (number of units of movement though
the input data)}
\usage{
tradeStatsRoll(Portfolios, Symbols, period, anchored = FALSE, k.span,
  k.step = 1, ...)
}
\description{
Produces 'rolling' statistics based on the 'period' (unit size),
span (input data window size), and step (number of units of movement though
the input data)
}

